TP_Facial_LandMarks/requirements.txt
mediapipe==0.10.21
opencv-python==4.11.0.86
streamlit==1.39.0
Pillow>=7.1.0,<11
numpy==1.26.4